Correctly anchor JSON brace-matching regular expression.
When large JSON payloads occur (e.g., in GNIP's main tweet feed), the
regular expression seems to not be quite as greedy as it should be,
instead matching a subset of the tweet data and therefore not having
the expected data.

By anchoring the regular expression to the start of the buffer
(allowing for some whitespace), the recursive routine's greediness
works as expected and the full JSON message is correctly interpreted.