Remove redundant code, rename remaining files and update README
As part of the final cleanup for this phase, we rename the `json_*`
files by removing that prefix as our gem now only deals with JSON.
The exception is `json_extractor` whose name is still relevant (its
function is to extract JSON objects from a text stream).

This allows us to use the `Stream` class directly, removing the need
for a `JsonStream` and `StreamDelegate`.

README updated to reflect the gem's scope.